requirements.txt
fastapi
uvicorn
python-docx
pdfplumber
openai
stripe
python-dotenv
markdown
jinja2
